Create PrintNightmare-Check
In July 2021, a zero-day exploit in Windows systems was released to the public detailing a method for remote code execution via a print spooler weakness. 
Microsoft quickly released a patch. TAMU UES IT needed to verify the upgrade of 113 company machines.
Purpose:
I wrote this script to verify the validity of updates on 113 company servers as part of PrinterNightmare incident response.