Create PanelVerification
TAMU UES IT manages the networking systems for a company within TAMU. The DNS/IP recording solution displayed errors for a great majority of this company's devices, meaning the information inputted in the solution was mostly incorrect. We needed to compare the DNS and IP information for a large number of devices in this network against the overarching TAMU records. This script does so through analysis of three Excel files that I collected.
Purpose:
An 83 line PowerShell script constructed to compare three excel documents containing over 73,000 server entries for differences in IP and DNS listings.